Move tests to test_runners.py
